+++ src/Spago/PackageSet.hs
@@ -151,11 +151,8 @@ updatePackageSetVersion maybeTag = do
             , directory = Dhall.Directory
               { components = [ currentTag, "download", "releases", repo, org ]}
             }
-          , ..
           }
-        , ..
         }
-      , ..
       } = [(org, repo, currentTag)]
     getCurrentTag _ = []
 
@@ -170,11 +167,9 @@ updatePackageSetVersion maybeTag = do
             { file = "packages.dhall"
             , directory = Dhall.Directory
               { components = [ _currentTag, "download", "releases", upgradeRepo, upgradeOrg ]}
-            , ..
             }
           , ..
           }
-        , ..
         }
       , ..
       }) | upgradeRepo == repo && upgradeOrg == org =
@@ -227,9 +222,7 @@ isRemoteFrozen (Dhall.Import
   { importHashed = Dhall.ImportHashed
     { importType = Dhall.Remote _
     , hash
-    , ..
     }
-  , ..
   })             = [isJust hash]
 isRemoteFrozen _ = []
 
@@ -276,7 +269,7 @@ freeze path = do
       (Dhall.PossiblyTransitiveInputFile path Dhall.NonTransitive)
       Dhall.Freeze.OnlyRemoteImports
       Dhall.Freeze.Secure
-      Dhall.Pretty.ASCII
+      (Just Dhall.Pretty.ASCII)
       Dhall.NoCensor