+++ src/Spago/Purs.hs
@@ -131,7 +131,7 @@ docs format sourcePaths = do
 pursVersion :: Text -> RIO env (Either Text Version.SemVer)
 pursVersion purs = Turtle.Bytes.shellStrictWithErr (purs <> " --version") empty >>= \case
   (ExitSuccess, out, _err) -> do
-    let versionText = headMay $ Text.split (== ' ') (Text.Encoding.decodeUtf8With lenientDecode out)
+    let versionText = headMay $ Text.split (== ' ') (Text.strip $ Text.Encoding.decodeUtf8With lenientDecode out)
         parsed = versionText >>= (hush . Version.semver)
 
     pure $ case parsed of